Create helper to sanitize colors
